checkpkg: Support for NeedFile()
A new style of dependency declaration.  Any check can simply call
NeedFile(filename, reason) which will cause checkpkg to look up owners
of these files and require them as dependencies.

If multiple files are listed with the same reason, any of the owners
will be sufficient.  When declaring reasons, it's important to include
some specific information (such as soname) in the reason.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@12032 d3b55034-1cff-0310-a425-aefe953e1e90