pkgdb: Hardlinks to packages in a catalog
This is the first step to catalog generation.  It creates hardlinks from
alpkgs to 'unstable'.  It does not generate the catalog nor description
files.

To get working catalogs, bldcat needs to be run in all generated
catalogs.

It works in a somewhat stupid way, because it unlinks all existing
package files in the catalogs.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@12975 d3b55034-1cff-0310-a425-aefe953e1e90