checkpkg: Allow to run more than 1 test per class
When writing tests for checkpkg checks, it would be nice to be able to
write more than one test in one class, instead of creating a new class
for every check.  Adding a level of indirection, a function with takes
a callback function as an argument, allows to do that.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13205 d3b55034-1cff-0310-a425-aefe953e1e90