csw-upload-pkg: OS-release-specific uploads
Update to csw-upload-pkg based on a suggestion from Peter Felecan.  When
uploading a 5.9 file, for example, it'll look at the current state of the
catalog, and if a 5.10-specific package is present, csw-upload-pkg will only
upload the 5.9 package to the 5.9 catalog.  If, on the other hand, there is
the same version of the package in 5.10 (and 5.11), csw-upload-pkg will upload
the package there as well.

There's a new flag, --os-release, which makes csw-upload-pkg upload a package
to one, specified, catalog only.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13307 d3b55034-1cff-0310-a425-aefe953e1e90