pkgdb-web: Return 'null' when package not found
This way, we can distinguish the case of a package not in catalog from a case
of using a wrong URL.

A JSON 'null' translates to Python None, which evaluates as boolean false.
This way we know there's no package.  When a HTTP 404 error is thrown, it
means that the URL used is wrong.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13334 d3b55034-1cff-0310-a425-aefe953e1e90