checkpkg: Disable a non-passing unit test (2)
When checkpkg is run twice against the same package, errors are
accumulating.  I'm not sure why.  At the moment, checkpkg never runs
checks more than once in one go, so it doesn't cause operational
problems, but this bug ought to be eventually fixed.  Disabling for now,
to work on something else.

Signed-off-by: Maciej Bliziński <maciej@opencsw.org>

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13878 d3b55034-1cff-0310-a425-aefe953e1e90