pkgdb-web: A bugfix for error tag display
The problem was that the function returning all error tags was asked for tags
matching arch='all', and no such tags ever exist: even if the package is
architecture-independent, checks are done always specific to an architecture
architecture and are stored as such.

When the package is arch='all', errors for all architectures are displayed.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@13899 d3b55034-1cff-0310-a425-aefe953e1e90