csw-upload-pkg: --no-filename-check
An option to skip the checking of filenames. The safeguard is intented do
catch situations such as a missing architecture. For example, if you upload
the following set of files:

foo-1.0-SunOS5.9-sparc-CSW.pkg.gz
foo-1.0-SunOS5.10-sparc-CSW.pkg.gz
foo-1.0-SunOS5.9-i386-CSW.pkg.gz

...the tool will tell you that foo-1.0-SunOS5.10-i386-CSW.pkg.gz is missing.

But sometimes there are legitimate irregularities, and this option allows to
do that.

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@15491 d3b55034-1cff-0310-a425-aefe953e1e90