safe_remove_package.py: Revive
The safe_remove_package utility was not working, because it was shelling out
to csw-upload-pkg which had the --remove option removed some 4 months ago.
Instead of calling csw-upload-pkg, it's not using the RestClient class from
rest.py. It's now much faster, because it's avoiding the csw-upload-pkg
startup time.

Happy removing!

git-svn-id: http://gar.svn.sourceforge.net/svnroot/gar/csw/mgar/gar/v2@19649 d3b55034-1cff-0310-a425-aefe953e1e90