pkgdb: A grand rewrite
https://sourceforge.net/apps/trac/gar/ticket/77

The starting problem was that the pkgstat blobs were too large, causing
timeouts and general unhappiness across the infrastructure. The main
pain point was that it was no longer possible to display package
metadata on the web interface.

Yak shaving ensued. I hereby present you with a half shaven yak.

- Elfdump information is stored in a separate database table, split by
  binary, indexed by binary's MD5 sum.
- Saving and retrieving JSON blobs is done over REST.
- Database registration is simplified. Unless a package is added to
  a catalog, packages have no entries in the relational part of the
  database.
- Package stats collection is done in a subprocess, allowing the OS to
  clean up all memory which is freed. This solves the memory
  fragmentation problem when indexing the whole catalog.