transform back symbol info from list to ElfSymbol namedtuple when elfdump data are retrieved
json doesn't preserve the namedtuple, they are translated to list.
So we must post process the json data to transform them back into
namedtuple.

The function was already present but not activated. This patch also
removes the function from ElfExtractor to put it directly in checkpkg_lib
to remove a dependency.